refactor(pdf): extract canonical quality analysis
